construction of GC should not fail for restmapper error caused by tpr
